Rename old StorageImpl to StorageWrapperImpl.h to avoid ambiguous
